Make snippets available when outside of Ruby
Closes https://github.com/JoshCheek/atom-seeing-is-believing/pull/26

These snippets are really like "start me out with a template", they
aren't intended to be used in the middle of an existing document. As
such, it's very common to try expanding them in a new window, before
setting the language to Ruby. There is no visual information that it's
wrong in this case (ie missing syntax highlighting), so it just looks
like SiB is broken.

I could blame the user, but this is bad UX on my part, I should treat
them like templates rather than snippets and just make them always
available.

The risk of collision is mitigated since the names are all likely to be
unique.

It would be really nice if, after expanding, we could set the language
to Ruby, just in case the user did do this, we now know that the lang
should be Ruby, so take care of that for them instead of requiring them
to realize the highlighting is off, figure out why and figure out how to
change it (eg a lot of users save the file because it has this same
filetype inference behaviour and they don't realize they can just set it
in the editor). This is the second time today I've wanted some behaviour
like this, I opened an issue requesting the ability to do it at
https://github.com/atom/snippets/issues/221