Proof of concept: Auto execute on "# => " snippet expansion
Needs to be cleaned up, this was just to see if it is possible, which it is.
I'm committing it on a branch to maybe prompt consideration in the snippets package.
It seems like maybe the way to do this for real would be for the snippets package
to emit an event for did-expand-snippet. Though I don't fully grasp how all the
pieces fit together.

For context, see:
  https://github.com/JoshCheek/atom-seeing-is-believing/issues/25
  https://github.com/atom/snippets/issues/221