Lets see if we can't get some tests on this thing!
Been avoiding it b/c it's all integration, so super expensive and
requires figuring out a lot about how Atom works. But feel like I have a
reasonable idea now, and it was super annoying to manually test while I
worked on that last set of changes. So, we'll give it a shot :)